kernel: add recursive_spinlock_t
