+++ moose/net/frame.c
@@ -4,7 +4,7 @@
 #include <mm/kmalloc.h>
 #include <net/frame.h>
 #include <net/inet.h>
-#include <sched/spinlock.h>
+#include <sched/locks.h>
 #include <string.h>
 
 #define FREE_FRAMES_COUNT 32
